+++ src/core/server/commands/commands.ts
@@ -1,4 +1,5 @@
 import './adminChat';
+import './audioStream';
 import './currency';
 import './death';
 import './getVehicle';