+++ src/core/shared/enums/system.ts
@@ -12,6 +12,7 @@ export enum SYSTEM_EVENTS {
     NOCLIP_RESET = 'noclip:Reset',
     //
     PLAYER_EMIT_ANIMATION = 'animation:Play',
+    PLAYER_EMIT_AUDIO_STREAM = 'audio:Stream',
     PLAYER_EMIT_SOUND_2D = 'sound:2D',
     PLAYER_EMIT_SOUND_3D = 'sound:3D',
     PLAYER_EMIT_FRONTEND_SOUND = 'sound:Frontend',